+++ ui/components/welcome-mat/WelcomeMatTile.jsx
@@ -1,27 +1,30 @@
 // Copyright (c) 2015-present, salesforce.com, inc. All rights reserved
 // Licensed under BSD 3-Clause - see LICENSE.txt or git.io/sfdc-license
 
-import React from 'react';
+import React, { Fragment } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
-import { VisualPickerMediaObject } from '../visual-picker/link/example';
 import { UtilityIcon } from '../icons/base/example';
 import { ActionIcon } from '../icons/action/example';
 
 class WelcomeMatTile extends React.Component {
   render() {
-    const { tile } = this.props;
+    const { tile, isInfoOnly } = this.props;
     const completed = tile.completed;
     const className = classNames('slds-welcome-mat__tile', {
-      'slds-welcome-mat__tile_complete': completed
+      'slds-welcome-mat__tile_complete': completed,
+      'slds-welcome-mat__tile_info-only': isInfoOnly
     });
 
-    return (
-      <li className={className}>
-        <VisualPickerMediaObject
-          symbol={tile.symbol}
-          icon={
+    const renderMediaObject = () => (
+      <Fragment>
+        <div
+          className={classNames(
+            'slds-media__figure slds-media__figure_fixed-width slds-align_absolute-center'
+          )}
+        >
+          <div className="slds-welcome-mat__tile-figure">
             <div className="slds-welcome-mat__tile-icon-container">
               <UtilityIcon
                 className="slds-icon-text-default"
@@ -29,20 +32,40 @@ class WelcomeMatTile extends React.Component {
                 title={false}
                 assistiveText={false}
               />
-              <ActionIcon
-                title="Completed"
-                assistiveText="Completed"
-                className="slds-welcome-mat__icon-check"
-                symbol="check"
-              />
+              {!isInfoOnly && (
+                <ActionIcon
+                  title="Completed"
+                  assistiveText="Completed"
+                  className="slds-welcome-mat__icon-check"
+                  symbol="check"
+                />
+              )}
             </div>
-          }
-        >
-          <h3 className="slds-welcome-mat__tile-title">{tile.title}</h3>
-          <p className="slds-welcome-mat__tile-description">
-            {tile.description}
-          </p>
-        </VisualPickerMediaObject>
+          </div>
+        </div>
+        <div className="slds-media__body">
+          <div className="slds-welcome-mat__tile-body">
+            <h3 className="slds-welcome-mat__tile-title">{tile.title}</h3>
+            <p className="slds-welcome-mat__tile-description">
+              {tile.description}
+            </p>
+          </div>
+        </div>
+      </Fragment>
+    );
+
+    return (
+      <li className={className}>
+        {isInfoOnly ? (
+          <div className="slds-media">{renderMediaObject()}</div>
+        ) : (
+          <a
+            href="javascript:void(0);"
+            className="slds-box slds-box_link slds-media"
+          >
+            {renderMediaObject()}
+          </a>
+        )}
       </li>
     );
   }