+++ ui/components/welcome-mat/__tests__/enzyme.spec.js
@@ -1,15 +1,16 @@
 // Copyright (c) 2015-present, salesforce.com, inc. All rights reserved
 // Licensed under BSD 3-Clause - see LICENSE.txt or git.io/sfdc-license
+
 /* eslint-env jest */
 
+import { mount, render, shallow } from 'enzyme';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { shallow, render, mount } from 'enzyme';
-
-import WelcomeMat, { sampleTiles } from '../';
-import WelcomeMatTile from '../WelcomeMatTile';
+import WelcomeMat from '../';
+import { sampleTiles, WelcomeMatMarketingContent } from '../helpers';
 import WelcomeMatContent from '../WelcomeMatContent';
 import WelcomeMatContentTrailhead from '../WelcomeMatContentTrailhead';
+import WelcomeMatTile from '../WelcomeMatTile';
 
 const tiles = sampleTiles();
 const twoTilesCompleted = [
@@ -18,7 +19,7 @@ const twoTilesCompleted = [
   ...tiles.slice(2)
 ];
 
-describe('Welcome Mat', () => {
+describe('Welcome Mat Component', () => {
   it('renders without crashing', () => {
     ReactDOM.render(<WelcomeMat />, document.createElement('div'));
   });
@@ -27,9 +28,38 @@ describe('Welcome Mat', () => {
     const rendered = render(<WelcomeMat tiles={twoTilesCompleted} />);
     expect(rendered.find('.slds-welcome-mat__tile_complete')).toHaveLength(2);
   });
+
+  it('correctly renders in info-only mode', () => {
+    const rendered = render(<WelcomeMat isInfoOnly />);
+    expect(
+      rendered.find('.slds-welcome-mat').hasClass('slds-welcome-mat_info-only')
+    ).toBe(true);
+    expect(
+      rendered
+        .find('.slds-welcome-mat__tiles')
+        .hasClass('slds-welcome-mat__tiles_info-only')
+    ).toBe(true);
+  });
+
+  it('correctly renders in splash mode', () => {
+    const rendered = render(
+      <WelcomeMat
+        tiles={null}
+        content={({ labelId }) => (
+          <WelcomeMatMarketingContent labelId={labelId} />
+        )}
+      />
+    );
+    expect(
+      rendered.find('.slds-welcome-mat').hasClass('slds-welcome-mat_splash')
+    ).toBe(true);
+    expect(
+      rendered.find('.slds-welcome-mat__info').hasClass('slds-size_1-of-1')
+    ).toBe(true);
+  });
 });
 
-describe('Welcome Mat Tile', () => {
+describe('Welcome Mat Tile Component', () => {
   it('renders a tile without crashing', () => {
     ReactDOM.render(<WelcomeMatTile />, document.createElement('div'));
   });
@@ -55,7 +85,7 @@ describe('Welcome Mat Tile', () => {
   });
 });
 
-describe('Welcome Mat Content', () => {
+describe('Welcome Mat Content Component', () => {
   const complete = 3;
   const total = 5;
   const completePercent = complete / total * 100;
@@ -80,7 +110,7 @@ describe('Welcome Mat Content', () => {
   });
 });
 
-describe('Welcome Mat Trailhead Content', () => {
+describe('Welcome Mat Trailhead Content Component', () => {
   const shallowNew = shallow(<WelcomeMatContentTrailhead />);
   const shallowComplete = shallow(
     <WelcomeMatContentTrailhead complete={5} total={5} />
@@ -118,5 +148,22 @@ describe('Welcome Mat Trailhead Content', () => {
   it('renders a brand button when complete', () => {
     expect(wrapper.find('.slds-button').exists()).toBe(true);
     expect(wrapper.find('.slds-button_brand').exists()).toBe(true);
+  });
+});
+
+describe('Welcome Mat / Modal Markup Separation', () => {
+  const wrapper = mount(<WelcomeMat />);
+
+  it('renders a welcome mat within modal content', () => {
+    expect(
+      wrapper.find('.slds-modal__content .slds-welcome-mat').exists()
+    ).toBe(true);
+    expect(
+      wrapper
+        .find(
+          '.slds-modal__content .slds-welcome-mat .slds-welcome-mat__content'
+        )
+        .exists()
+    ).toBe(true);
   });
 });